requirements.txt
libg3n
pyyaml